Added very light weight parser for simplex trees
Implementation in lwtree.py is extremely barebones, but around 10 times faster according to my timing.
Its tree implementation does not offer much functionality and is admittedly not exactly intuitive to use, but can be reliably used for the most relevant operations, which is discovering simplices, maximal simplices and counting them.

Implementations of these most common operations are also provided